+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/search/CertifiedProductSearchResult.java
@@ -2,26 +2,75 @@ package gov.healthit.chpl.domain.search;
 
 import java.io.Serializable;
 
+import com.fasterxml.jackson.annotation.JsonView;
+
 public class CertifiedProductSearchResult implements Serializable {
 	private static final long serialVersionUID = -2547390525592841034L;
-	private Long id;
-	private String chplProductNumber;
-	private String edition;
-	private String atl;
-	private String acb;
-	private String acbCertificationId;
-	private String practiceType;
-	private String developer;
-	private String product;
-	private String version;
-	private Long certificationDate;
-	private String certificationStatus;
-	private boolean hasOpenSurveillance;
-	private boolean hasClosedSurveillance;
-	private boolean hasOpenNonconformities;
-	private boolean hasClosedNonconformities;
+	
+	@JsonView({SearchViews.Default.class})
+	protected Long id;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String chplProductNumber;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String edition;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String atl;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String acb;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String acbCertificationId;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String practiceType;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String developer;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String product;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String version;
+	
+	@JsonView({SearchViews.Default.class})
+	protected Long certificationDate;
+	
+	@JsonView({SearchViews.Default.class})
+	protected String certificationStatus;
+	
+	@JsonView({SearchViews.Default.class})
+	protected Long surveillanceCount;
+	
+	@JsonView({SearchViews.Default.class})
+	protected Long openNonconformityCount;
+	
+	@JsonView({SearchViews.Default.class})
+	protected Long closedNonconformityCount;
 	
 	public CertifiedProductSearchResult() {
+	}
+	
+	public CertifiedProductSearchResult(CertifiedProductSearchResult other) {
+		this.id = other.getId();
+		this.chplProductNumber = other.getChplProductNumber();
+		this.edition = other.getEdition();
+		this.atl = other.getAtl();
+		this.acb = other.getAcb();
+		this.acbCertificationId = other.getAcbCertificationId();
+		this.practiceType = other.getPracticeType();
+		this.developer = other.getDeveloper();
+		this.product = other.getProduct();
+		this.version = other.getVersion();
+		this.certificationDate = other.getCertificationDate();
+		this.certificationStatus = other.getCertificationStatus();
+		this.surveillanceCount = other.getSurveillanceCount();
+		this.openNonconformityCount = other.getOpenNonconformityCount();
+		this.closedNonconformityCount = other.getClosedNonconformityCount();
 	}
 	
 	public Long getId() {
@@ -90,17 +139,24 @@ public class CertifiedProductSearchResult implements Serializable {
 	public void setCertificationStatus(String certificationStatus) {
 		this.certificationStatus = certificationStatus;
 	}
-	public boolean isHasOpenSurveillance() {
-		return hasOpenSurveillance;
+	public Long getSurveillanceCount() {
+		return surveillanceCount;
 	}
-	public void setHasOpenSurveillance(boolean hasOpenSurveillance) {
-		this.hasOpenSurveillance = hasOpenSurveillance;
+	public void setSurveillanceCount(Long surveillanceCount) {
+		this.surveillanceCount = surveillanceCount;
 	}
-	public boolean isHasOpenNonconformities() {
-		return hasOpenNonconformities;
+	public Long getOpenNonconformityCount() {
+		return openNonconformityCount;
 	}
-	public void setHasOpenNonconformities(boolean hasOpenNonconformities) {
-		this.hasOpenNonconformities = hasOpenNonconformities;
+	public void setOpenNonconformityCount(Long openNonconformityCount) {
+		this.openNonconformityCount = openNonconformityCount;
+	}
+
+	public Long getClosedNonconformityCount() {
+		return closedNonconformityCount;
+	}
+	public void setClosedNonconformityCount(Long closedNonconformityCount) {
+		this.closedNonconformityCount = closedNonconformityCount;
 	}
 	
 	public String getAcbCertificationId() {
@@ -109,21 +165,5 @@ public class CertifiedProductSearchResult implements Serializable {
 
 	public void setAcbCertificationId(String acbCertificationId) {
 		this.acbCertificationId = acbCertificationId;
-	}
-
-	public boolean isHasClosedSurveillance() {
-		return hasClosedSurveillance;
-	}
-
-	public void setHasClosedSurveillance(boolean hasClosedSurveillance) {
-		this.hasClosedSurveillance = hasClosedSurveillance;
-	}
-
-	public boolean isHasClosedNonconformities() {
-		return hasClosedNonconformities;
-	}
-
-	public void setHasClosedNonconformities(boolean hasClosedNonconformities) {
-		this.hasClosedNonconformities = hasClosedNonconformities;
 	}
 }