+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/CertifiedProductSearchManager.java
@@ -1,12 +1,15 @@
 package gov.healthit.chpl.manager;
 
+import java.util.List;
+
 import gov.healthit.chpl.domain.SearchRequest;
 import gov.healthit.chpl.domain.SearchResponse;
 import gov.healthit.chpl.domain.search.BasicSearchResponse;
+import gov.healthit.chpl.domain.search.CertifiedProductFlatSearchResult;
 
 
 public interface CertifiedProductSearchManager {
 
-	public BasicSearchResponse search();
+	public List<CertifiedProductFlatSearchResult> search();
 	public SearchResponse search(SearchRequest searchRequest);
 }